Edit Gallery name and adding update buttons
Added a lot of functionality today. Added an edit gallery option to each gallery which redirects to a new page with a form that allows the name of the gallery to be updated.

Also added a button on the show.erb page which deletes the gallery entirely.  It then re-directs to the index where you can see the gallery is no longer displayed.