Added edit images feature to app.rb
Created a new page to add a new images to a gallery.  This meant I had to created a new html page just to handle new requests. I also had to crate a form within that html view page with an embeded form.  The form passed the key value pair of image and image url.  Then I created a new route in app.rb to handle the post request.  It took the hash of the user params and merged with another hash of the gallery_id and params(gallery_id).